move commands, add missing docs.
