+++ internal/command/command.go
@@ -1,3 +1,5 @@
+// Package commands provides the interfaces for commands that can be pushed
+// into a command queue for the controller to consume.
 package command
 
 import "github.com/martinohmann/rfoutlet/internal/outlet"
@@ -18,6 +20,21 @@ type Command interface {
 	// Must return true if a potential state change should be broadcasted to
 	// all connected clients.
 	Execute(context Context) (broadcast bool, err error)
+}
+
+// Sender can send messages.
+type Sender interface {
+	// Send sends out a message.
+	Send(msg []byte)
+}
+
+// SenderAwareCommand is aware of the comand's sender.
+type SenderAwareCommand interface {
+	Command
+
+	// SetSender sets the sender on the command. The sender can be used to send
+	// messages back to the client that issued the command.
+	SetSender(sender Sender)
 }
 
 // NewTestContext creates a new Context which can be used in tests. It returns
