+++ internal/controller/commands/commands_test.go
@@ -1,6 +1,8 @@
-package websocket
+package commands
 
 import (
+	"bytes"
+	"errors"
 	"testing"
 
 	"github.com/martinohmann/rfoutlet/internal/command"
@@ -9,6 +11,32 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
+
+type fakeSender struct {
+	buf bytes.Buffer
+}
+
+func (s *fakeSender) Send(msg []byte) {
+	s.buf.Write(msg)
+}
+
+func TestStatusCommand(t *testing.T) {
+	ctx, r, _ := command.NewTestContext()
+
+	s := &fakeSender{}
+
+	r.RegisterGroups(&outlet.Group{ID: "foo"})
+
+	cmd := StatusCommand{}
+	cmd.SetSender(s)
+
+	broadcast, err := cmd.Execute(ctx)
+
+	require.NoError(t, err)
+	assert.False(t, broadcast)
+
+	assert.Equal(t, `[{"id":"foo","displayName":"","outlets":null}]`, s.buf.String())
+}
 
 func TestOutletCommand(t *testing.T) {
 	ctx, r, _ := command.NewTestContext()
@@ -90,4 +118,58 @@ func TestIntervalCommand(t *testing.T) {
 
 	require.NoError(t, err)
 	assert.True(t, broadcast)
+}
+
+func TestStateCorrectionCommand(t *testing.T) {
+	tests := []struct {
+		name              string
+		outletState       outlet.State
+		desiredState      outlet.State
+		switchErr         error
+		expectedState     outlet.State
+		expectedBroadcast bool
+		expectedErr       error
+	}{
+		{
+			name:          "outlet is already in desired state",
+			outletState:   outlet.StateOn,
+			desiredState:  outlet.StateOn,
+			expectedState: outlet.StateOn,
+		},
+		{
+			name:              "outlet is not in desired state",
+			outletState:       outlet.StateOn,
+			desiredState:      outlet.StateOff,
+			expectedState:     outlet.StateOff,
+			expectedBroadcast: true,
+		},
+		{
+			name:              "outlet is not in desired state, switch error",
+			outletState:       outlet.StateOn,
+			desiredState:      outlet.StateOff,
+			switchErr:         errors.New("whoops"),
+			expectedState:     outlet.StateOn,
+			expectedBroadcast: false,
+			expectedErr:       errors.New("whoops"),
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			o := &outlet.Outlet{State: test.outletState}
+			cmd := StateCorrectionCommand{
+				Outlet:       o,
+				DesiredState: test.desiredState,
+			}
+
+			ctx, _, s := command.NewTestContext()
+			s.Err = test.switchErr
+
+			broadcast, err := cmd.Execute(ctx)
+
+			assert.Equal(t, test.expectedErr, err)
+			assert.Equal(t, test.expectedBroadcast, broadcast)
+			assert.Equal(t, test.expectedState, o.GetState())
+		})
+	}
 }
