+++ internal/schedule/schedule.go
@@ -1,3 +1,5 @@
+// Package schedule provides types to define time switch schedule intervals for
+// outlets.
 package schedule
 
 import (
@@ -9,25 +11,25 @@ import (
 	uuid "github.com/satori/go.uuid"
 )
 
-// Schedule type definition
+// Schedule is a collection of intervals.
 type Schedule struct {
 	sync.RWMutex
 	intervals []Interval
 }
 
-// New create a new schedule
+// New creates a new empty *Schedule.
 func New() *Schedule {
 	return NewWithIntervals(make([]Interval, 0))
 }
 
-// NewWithIntervals create a new schedule seeded with intervals
+// NewWithIntervals create a new *Schedule with intervals.
 func NewWithIntervals(intervals []Interval) *Schedule {
 	return &Schedule{
 		intervals: intervals,
 	}
 }
 
-// Enabled returns true if any of the intervals is enabled
+// Enabled returns true if any of the intervals is enabled.
 func (s *Schedule) Enabled() bool {
 	if s == nil {
 		return false
@@ -46,7 +48,7 @@ func (s *Schedule) Enabled() bool {
 	return false
 }
 
-// Contains returns true if any of the intervals contains t
+// Contains returns true if any of the intervals contains t.
 func (s *Schedule) Contains(t time.Time) bool {
 	if s == nil {
 		return false
@@ -65,7 +67,7 @@ func (s *Schedule) Contains(t time.Time) bool {
 	return false
 }
 
-// AddInterval adds an interval to the schedule of an outlet
+// AddInterval adds an interval to the schedule of an outlet.
 func (s *Schedule) AddInterval(interval Interval) error {
 	if interval.ID == "" {
 		interval.ID = uuid.NewV4().String()
@@ -117,12 +119,18 @@ func (s *Schedule) DeleteInterval(interval Interval) error {
 	return fmt.Errorf("interval %q does not exist", interval.ID)
 }
 
-// UnmarshalJSON implements the json.Unmarshaler interface
+// UnmarshalJSON implements the json.Unmarshaler interface.
+//
+// This ensures that the json bytes are correctly unmarshalled into the
+// internal slice of Interval values.
 func (s *Schedule) UnmarshalJSON(b []byte) error {
 	return json.Unmarshal(b, &s.intervals)
 }
 
-// MarshalJSON implements the json.Marshaler interface
+// MarshalJSON implements the json.Marshaler interface.
+//
+// This hides that fact that *Schedule wraps a slice of Interval in the
+// marshalled json.
 func (s *Schedule) MarshalJSON() ([]byte, error) {
 	return json.Marshal(s.intervals)
 }
