+++ internal/timeswitch/timeswitch_test.go
@@ -6,6 +6,7 @@ import (
 	"time"
 
 	"github.com/martinohmann/rfoutlet/internal/command"
+	"github.com/martinohmann/rfoutlet/internal/controller/commands"
 	"github.com/martinohmann/rfoutlet/internal/outlet"
 	"github.com/martinohmann/rfoutlet/internal/schedule"
 	"github.com/stretchr/testify/assert"
@@ -60,7 +61,7 @@ func TestTimeSwitch(t *testing.T) {
 				},
 			},
 			expectedCommands: []command.Command{
-				TimeSwitchCommand{
+				commands.StateCorrectionCommand{
 					DesiredState: outlet.StateOn,
 					Outlet: &outlet.Outlet{
 						State: outlet.StateOff,
@@ -108,7 +109,7 @@ func TestTimeSwitch(t *testing.T) {
 				},
 			},
 			expectedCommands: []command.Command{
-				TimeSwitchCommand{
+				commands.StateCorrectionCommand{
 					DesiredState: outlet.StateOff,
 					Outlet: &outlet.Outlet{
 						State: outlet.StateOn,
